backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
supabase>=2.5.0
google-genai>=0.3.0
sentence-transformers>=3.0.0
yt-dlp>=2024.4.9
instaloader>=4.11
python-multipart>=0.0.9
